integrate "fast" shack hartmann and lineofsight into main code
