+++ soapy/lineofsight.py
@@ -25,11 +25,11 @@ Examples::
 """
 
 import numpy
-from scipy.interpolate import interp2d
-
-from . import logger, interp
 
 from aotools import opticalpropagation
+
+from . import logger, interp
+from . import numbalib
 
 DTYPE = numpy.float32
 CDTYPE = numpy.complex64
@@ -58,8 +58,8 @@ class LineOfSight(object):
     """
     def __init__(
             self, config, soapyConfig,
-            propagationDirection="down", out_pixel_scale=None,
-            nx_out_pxls=None, mask=None, metaPupilPos=None):
+            propagation_direction="down", out_pixel_scale=None,
+            nx_out_pixels=None, mask=None, metaPupilPos=None):
 
         self.config = config
         self.soapy_config = soapyConfig
@@ -69,26 +69,38 @@ class LineOfSight(object):
         self.wavelength = self.config.wavelength
         self.telescope_diameter = self.soapy_config.tel.telDiam
 
+        self.source_altitude = self.height
+
+        self.nx_scrn_size = self.soapy_config.sim.scrnSize
+        self.n_layers = self.soapy_config.atmos.scrnNo
+        self.layer_altitudes = self.soapy_config.atmos.scrnHeights
+
+        self.n_dm = self.soapy_config.sim.nDM
+        self.dm_altitudes = numpy.array([self.soapy_config.dms[dm_n].altitude for dm_n in range(self.n_dm)])
+
+        # If the line of sight has a launch position, must include in calculations! Conver to metres from centre of telescope
+        try:
+            self.launch_position = self.config.launchPosition * self.telescope_diameter/2.
+        except AttributeError:
+            self.launch_position = numpy.array([0, 0])
+
+        self.threads = self.soapy_config.sim.threads
+
         self.simConfig = soapyConfig.sim
         self.atmosConfig = soapyConfig.atmos
-        self.soapyConfig = soapyConfig
 
         self.mask = mask
+        self.propagation_direction = propagation_direction
 
-        self.calcInitParams(out_pixel_scale, nx_out_pxls)
-
-        self.propagationDirection = propagationDirection
-
-        # If GS not at infinity, find meta-pupil radii for each layer
-        if self.height!=0:
-            self.radii = self.findMetaPupilSizes(self.height)
-        else:
-            self.radii = None
+        self.calcInitParams(out_pixel_scale, nx_out_pixels)
 
         self.allocDataArrays()
 
         # Can be set to use other values as metapupil position
         self.metaPupilPos = metaPupilPos
+
+        self.thread_pool = numbalib.ThreadPool(self.threads)
+
 
     # Some attributes for compatability between WFS and others
     @property
@@ -123,23 +135,25 @@ class LineOfSight(object):
             self.config.GSPosition
             self.config.GSPosition = position
 
-
 ############################################################
 # Initialisation routines
+
+
     def calcInitParams(self, out_pixel_scale=None, nx_out_pixels=None):
         """
         Calculates some parameters required later
 
         Parameters:
-            out_pixel_scale (float): Pixel scale of required phase/EField (metres/pxl)
-            nx_out_pixels (int): Size of output array in pixels
+            outPxlScale (float): Pixel scale of required phase/EField (metres/pxl)
+            nOutPxls (int): Size of output array in pixels
         """
+        logger.debug("Calculate LOS Init PArams!")
         # Convert phase deviation to radians at wfs wavelength.
         # (currently in nm remember...?)
         self.phs2Rad = 2*numpy.pi/(self.wavelength * 10**9)
 
         # Get the size of the phase required by the system
-        self.inPxlScale = self.phase_pixel_scale
+        self.in_pixel_scale = self.phase_pixel_scale
 
         if out_pixel_scale is None:
             self.out_pixel_scale = self.phase_pixel_scale
@@ -154,7 +168,65 @@ class LineOfSight(object):
         if self.mask is not None:
             self.outMask = interp.zoom(
                     self.mask, self.nx_out_pixels).round()
+        else:
+            self.outMask = None
 
+        self.output_phase_diameter = self.nx_out_pixels * self.out_pixel_scale
+
+        # Calculate coords of phase at each altitude
+        self.layer_metapupil_coords = numpy.zeros((self.n_layers, 2, self.nx_out_pixels))
+        for i in range(self.n_layers):
+            x1, x2, y1, y2 = self.calculate_altitude_coords(self.layer_altitudes[i])
+            self.layer_metapupil_coords[i, 0] = numpy.linspace(x1, x2-1, self.nx_out_pixels)
+            self.layer_metapupil_coords[i, 1] = numpy.linspace(y1, y2-1, self.nx_out_pixels)
+
+        # ensure coordinates aren't out of bounds for interpolation
+        self.layer_metapupil_coords = self.layer_metapupil_coords.clip(0, self.nx_scrn_size - 1.000000001)
+
+        # Calculate coords of phase at each DM altitude
+        self.dm_metapupil_coords = numpy.zeros((self.n_dm, 2, self.nx_out_pixels))
+        for i in range(self.n_dm):
+            x1, x2, y1, y2 = self.calculate_altitude_coords(self.dm_altitudes[i])
+            self.dm_metapupil_coords[i, 0] = numpy.linspace(x1, x2-1, self.nx_out_pixels)
+            self.dm_metapupil_coords[i, 1] = numpy.linspace(y1, y2-1, self.nx_out_pixels)
+        self.dm_metapupil_coords = self.dm_metapupil_coords.clip(0, self.nx_scrn_size - 1.000000001)
+
+        self.radii = None
+
+        self.phase_screens = numpy.zeros((self.n_layers, self.nx_out_pixels, self.nx_out_pixels))
+        self.correction_screens = numpy.zeros((self.n_dm, self.nx_out_pixels, self.nx_out_pixels))
+        self.phase_correction = numpy.zeros((self.nx_out_pixels, self.nx_out_pixels))
+
+    def calculate_altitude_coords(self, layer_altitude):
+        """
+        Calculate the co-ordinates of vertices of fo the meta-pupil at altitude given a guide star
+        direction and source altitude
+
+        Paramters:
+            layer_altitude (float): Altitude of phase layer
+        """
+        direction_radians = ASEC2RAD * numpy.array(self.position)
+
+        centre = (direction_radians * layer_altitude)
+
+        # If propagating up must account for launch position
+        if self.propagation_direction == "up":
+            centre += self.launch_position * (1 - layer_altitude/self.source_altitude)
+
+        if self.source_altitude != 0:
+            meta_pupil_size = self.output_phase_diameter * (1 - layer_altitude / self.source_altitude)
+        else:
+            meta_pupil_size = self.output_phase_diameter
+
+        x1 = ((centre[0] - meta_pupil_size / 2.) / self.in_pixel_scale) + self.nx_scrn_size / 2.
+        x2 = ((centre[0] + meta_pupil_size / 2.) / self.in_pixel_scale) + self.nx_scrn_size / 2.
+        y1 = ((centre[1] - meta_pupil_size / 2.) / self.in_pixel_scale) + self.nx_scrn_size / 2.
+        y2 = ((centre[1] + meta_pupil_size / 2.) / self.in_pixel_scale) + self.nx_scrn_size / 2.
+
+        logger.debug("Layer Altitude: {}".format(layer_altitude))
+        logger.debug("Coords: x1: {}, x2: {}, y1: {},  y2: {}".format(x1, x2, y1, y2))
+
+        return x1, x2, y1, y2
 
     def allocDataArrays(self):
         """
@@ -169,143 +241,6 @@ class LineOfSight(object):
         self.EField = numpy.ones([self.nx_out_pixels] * 2, dtype=CDTYPE)
 
 
-    def findMetaPupilSizes(self, GSHeight):
-        '''
-        Evaluates the sizes of the effective metePupils
-        at each screen height - if a GS of finite height is used.
-
-        Parameters:
-            GSHeight (float): The height of the GS in metres
-
-        Returns:
-            dict : A dictionary containing the radii of a meta-pupil at each screen height in phase pixels
-        '''
-
-        radii = {}
-        for i in xrange(self.atmosConfig.scrnNo):
-            radii[i] = self.calcMetaPupilSize(
-                        self.atmosConfig.scrnHeights[i], GSHeight) / self.phase_pixel_scale
-
-        return radii
-
-
-    def calcMetaPupilSize(self, scrnHeight, GSHeight):
-        """
-        Calculates the radius of a meta pupil at a altitude layer, of a GS at a give altitude
-
-        Parameters:
-            scrnHeight (float): Altitude of meta-pupil
-            GSHeight (float): Altitude of guide star
-
-        Returns:
-            float: Radius of metapupil in metres
-
-        """
-        # If GS at infinity, radius is telescope radius
-        if self.height == 0:
-            return self.telescope_diameter/2.
-
-        # If scrn is above LGS, radius is 0
-        if scrnHeight >= GSHeight:
-            return 0
-
-        # Find radius of metaPupil geometrically (fraction of pupil at
-        # Ground Layer)
-        radius = (self.telescope_diameter/2.) * (1-(float(scrnHeight)/GSHeight))
-
-        return radius
-
-    #############################################################
-    # Phase stacking routines for a WFS frame
-    def getMetaPupilPos(self, height, apos=None):
-        '''
-        Finds the centre of a metapupil at a given height,
-        when offset by a given angle in arsecs, in metres from the central position
-
-        Parameters:
-            height (float): Height of the layer in metres
-            apos (ndarray, optional):  The angular position of the GS in asec. If not set, will use the config position
-
-        Returns:
-            ndarray: The position of the centre of the metapupil in metres
-        '''
-        # if no pos given, use system pos and convert into radians
-        if not numpy.any(apos):
-            pos = (numpy.array(self.position)).astype('float')
-
-        pos *= ASEC2RAD
-
-        # Position of centre of GS metapupil off axis at required height
-        GSCent = (numpy.tan(pos) * height)
-
-        return GSCent
-
-    def getMetaPupilPhase(
-            self, scrn, height, radius=None,  apos=None, pos=None):
-        '''
-        Returns the phase across a metaPupil at some height and angular
-        offset in arcsec. Interpolates phase to size of the pupil if cone
-        effect is required
-
-        Parameters:
-            scrn (ndarray): An array representing the phase screen
-            height (float): Height of the phase screen
-            radius (float, optional): Radius of the meta-pupil. If not set, will use system pupil size.
-            apos (ndarray, optional): X, Y angular position of the guide star in asecs, otherwise will use that set in config or 'pos'
-            pos (ndarray, optional): X, Y central position of the metapupil in metres. If None, then config used to calculate it from config pos, or 'apos'.
-
-        Return:
-            ndarray: The meta pupil at the specified height
-        '''
-
-        # If the radius is 0, then 0 phase is returned
-        if radius == 0:
-            return numpy.zeros((self.sim_size, self.sim_size))
-
-        if apos is not None:
-            GSCent = self.getMetaPupilPos(height, apos) / self.phase_pixel_scale
-        elif pos is not None:
-            GSCent = pos / self.phase_pixel_scale
-        else:
-            GSCent = self.getMetaPupilPos(height) / self.phase_pixel_scale
-
-        # Find the size in metres of phase that is required
-        self.phaseRadius = (self.out_pixel_scale * self.nx_out_pixels) / 2.
-        # And a corresponding coordinate on the phase screen
-        self.phaseCoord = int(round(self.phaseRadius / self.phase_pixel_scale))
-
-        logger.debug('phaseCoord:{}'.format(self.phaseCoord))
-        # The sizes of the phase screen
-        scrnX, scrnY = scrn.shape
-
-        # If the GS is not at infinity, take into account cone effect
-        if radius != None:
-            fact = float(2*radius)/self.pupil_size
-        else:
-            fact = 1
-
-        simSize = self.simConfig.simSize
-        x1 = scrnX/2. + GSCent[0] - fact * self.phaseCoord
-        x2 = scrnX/2. + GSCent[0] + fact * self.phaseCoord
-        y1 = scrnY/2. + GSCent[1] - fact * self.phaseCoord
-        y2 = scrnY/2. + GSCent[1] + fact * self.phaseCoord
-
-        logger.debug("LoS Scrn Coords - ({0}:{1}, {2}:{3})".format(
-                x1,x2,y1,y2))
-        if ( x1 < 0 or x2 > scrnX or y1 < 0 or y2 > scrnY):
-            logger.warning("GS separation requires larger screen size. \nheight: {3}, GSCent: {0}, \nscrnSize: {1}, phaseCoord, {8}, simSize: {2}, fact: {9}\nx1: {4},x2: {5}, y1: {6}, y2: {7}".format(
-                    GSCent, scrn.shape, simSize, height, x1, x2, y1, y2, self.phaseCoord, fact))
-            raise ValueError("Requested phase exceeds phase screen size. See log warnings.")
-
-        xCoords = numpy.linspace(x1, x2 - 1, self.nx_out_pixels)
-        yCoords = numpy.linspace(y1, y2 - 1, self.nx_out_pixels)
-        scrnCoords = numpy.arange(scrnX)
-        interpObj = interp2d(
-                scrnCoords, scrnCoords, scrn, copy=False)
-        metaPupil = interpObj(xCoords, yCoords)
-
-        self.metaPupil = metaPupil
-        return metaPupil
 ######################################################
 
     def zeroData(self, **kwargs):
@@ -324,11 +259,18 @@ class LineOfSight(object):
             radii (dict, optional): Radii of each meta pupil of each screen height in pixels. If not given uses pupil radius.
             apos (ndarray, optional):  The angular position of the GS in radians. If not set, will use the config position
         """
+        for i in range(self.scrns.shape[0]):
+            numbalib.bilinear_interp(
+                self.scrns[i], self.layer_metapupil_coords[i, 0], self.layer_metapupil_coords[i, 1],
+                self.phase_screens[i], self.thread_pool, bounds_check=False)
+
         # Check if geometric or physical
         if self.config.propagationMode == "Physical":
             return self.makePhasePhys(radii)
         else:
             return self.makePhaseGeometric(radii)
+
+
 
     def makePhaseGeometric(self, radii=None, apos=None):
         '''
@@ -339,29 +281,14 @@ class LineOfSight(object):
             apos (ndarray, optional):  The angular position of the GS in radians. If not set, will use the config position
         '''
 
-        for i in range(len(self.scrns)):
-            logger.debug("Layer: {}".format(i))
-            if radii is None:
-                radius = None
-            else:
-                radius = radii[i]
 
-            if self.metaPupilPos is None:
-                pos = None
-            else:
-                pos = self.metaPupilPos[i]
-
-            phase = self.getMetaPupilPhase(
-                    self.scrns[i], self.atmosConfig.scrnHeights[i],
-                    pos=pos, radius=radius)
-
-            self.phase += phase
+        self.phase_screens.sum(0, out=self.phase)
 
         # Convert phase to radians
         self.phase *= self.phs2Rad
 
         # Change sign if propagating up
-        if self.propagationDirection == 'up':
+        if self.propagation_direction == 'up':
             self.phase *= -1
 
         self.EField[:] = numpy.exp(1j*self.phase)
@@ -377,83 +304,10 @@ class LineOfSight(object):
             apos (ndarray, optional):  The angular position of the GS in radians. If not set, will use the config position
         '''
 
-        scrnNo = len(self.scrns)
-        z_total = 0
-        scrnRange = range(0, scrnNo)
-
-        # Get initial up/down dependent params
-        if self.propagationDirection == "up":
-            ht = 0
-            ht_final = self.config.height
-            if ht_final==0:
-                raise ValueError("Can't propagate up to infinity")
-            scrnAlts = self.atmosConfig.scrnHeights
-
-            self.EFieldBuf = self.outMask.copy().astype(CDTYPE)
-            logger.debug("Create EField Buf of mask")
-
-        else:
-            ht = self.atmosConfig.scrnHeights[scrnNo-1]
-            ht_final = 0
-            scrnRange = scrnRange[::-1]
-            scrnAlts = self.atmosConfig.scrnHeights[::-1]
-            self.EFieldBuf = numpy.exp(
-                    1j*numpy.zeros((self.nx_out_pixels,) * 2)).astype(CDTYPE)
-            logger.debug("Create EField Buf of zero phase")
-
-        # Propagate to first phase screen (if not already there)
-        if ht!=scrnAlts[0]:
-            logger.debug("propagate to first phase screen")
-            z = abs(scrnAlts[0] - ht)
-            self.EFieldBuf[:] = opticalpropagation.angularSpectrum(
-                        self.EFieldBuf, self.wavelength,
-                        self.out_pixel_scale, self.out_pixel_scale, z)
-
-        # Go through and propagate between phase screens
-        for i in scrnRange:
-            # Check optional radii and position
-            if radii is None:
-                radius = None
-            else:
-                radius = radii[i]
-
-            if self.metaPupilPos is None:
-                pos = None
-            else:
-                pos = self.metaPupilPos[i]
-
-            # Get phase for this layer
-            phase = self.getMetaPupilPhase(
-                    self.scrns[i],
-                    self.atmosConfig.scrnHeights[i], radius=radius, pos=pos)
-
-            # Convert phase to radians
-            phase *= self.phs2Rad
-
-            # Change sign if propagating up
-            if self.propagationDirection == 'up':
-                self.phase *= -1
-
-            # Get propagation distance for this layer
-            if i==(scrnNo-1):
-                z = abs(ht_final - ht) - z_total
-            else:
-                z = abs(scrnAlts[i+1] - scrnAlts[i])
-
-            # Update total distance counter
-            z_total += z
-
-            # Apply phase to EField
-            self.EFieldBuf *= numpy.exp(1j*phase)
-
-            # Do ASP for last layer to next
-            self.EFieldBuf[:] = opticalpropagation.angularSpectrum(
-                        self.EFieldBuf, self.wavelength,
-                        self.out_pixel_scale, self.out_pixel_scale, z)
-
-            logger.debug("Propagation: {}, {} m. Total: {}".format(i, z, z_total))
-
-            self.EField[:] = self.EFieldBuf
+        self.EField[:] = physical_atmosphere_propagation(
+            self.phase_screens, self.outMask, self.layer_altitudes, self.source_altitude,
+            self.wavelength, self.out_pixel_scale,
+            propagation_direction=self.propagation_direction)
 
         return self.EField
 
@@ -464,31 +318,21 @@ class LineOfSight(object):
         Parameters:
             correction (list or ndarray): either 2-d array describing correction, or list of correction arrays
         """
-        # If just an arary, put in list
-        # if isinstance(correction, numpy.ndarray):
-        #     correction = [correction]
-        
-        for corr in correction:
-            # If correction is a standard ndarray, assume at ground
-            if hasattr(corr, "altitude"):
-                altitude = corr.altitude
-            else:
-                altitude = 0
-                   
-            # Cut out the bit of the correction we need
-            metaPupilRadius = self.calcMetaPupilSize(
-                        altitude, self.height) / self.phase_pixel_scale
-            corr = self.getMetaPupilPhase(corr, altitude, radius=metaPupilRadius)
-            
-            # Correct EField
-            self.EField *= numpy.exp(-1j * corr * self.phs2Rad)
-            
-            # self.phase -= corr * self.phs2Rad
-            
-            # Also correct phase in case its required
-            self.residual = self.phase/self.phs2Rad - corr
-           
-            self.phase = self.residual * self.phs2Rad
+        for i in range(correction.shape[0]):
+            numbalib.bilinear_interp(
+                correction[i], self.dm_metapupil_coords[i, 0], self.dm_metapupil_coords[i, 1],
+                self.correction_screens[i], self.thread_pool, bounds_check=False)
+
+        self.correction_screens.sum(0, out=self.phase_correction)
+
+        # Correct EField
+        self.EField *= numpy.exp(-1j * self.phase_correction * self.phs2Rad)
+
+        # Also correct phase in case its required
+        self.residual = self.phase / self.phs2Rad - self.phase_correction
+
+        self.phase = self.residual * self.phs2Rad
+
 
     def frame(self, scrns=None, correction=None):
         '''
@@ -510,20 +354,97 @@ class LineOfSight(object):
         self.zeroData()
 
         if scrns is not None:
-
-            #If scrns is not dict or list, assume array and put in list
-            # t = type(scrns)
-            # if t != dict and t != list:
-            #     scrns = [scrns]
-
-            if scrns.ndim == 2:
-                scrns = scrns.reshape((1, scrns.shape[0], scrns.shape[1]))
+            if scrns.ndim==2:
+                scrns.shape = 1, scrns.shape[0], scrns.shape[1]
             self.scrns = scrns
-
             self.makePhase(self.radii)
-        
+
         self.residual = self.phase        
         if correction is not None:
             self.performCorrection(correction)
 
         return self.residual
+
+
+def physical_atmosphere_propagation(
+            phase_screens, output_mask, layer_altitudes, source_altitude,
+            wavelength, output_pixel_scale,
+            propagation_direction="up"):
+    '''
+    Finds total line of sight complex amplitude by propagating light through phase screens
+
+    Parameters:
+        radii (dict, optional): Radii of each meta pupil of each screen height in pixels. If not given uses pupil radius.
+        apos (ndarray, optional):  The angular position of the GS in radians. If not set, will use the config position
+    '''
+
+    scrnNo = len(phase_screens)
+    z_total = 0
+    scrnRange = range(0, scrnNo)
+
+    nx_output_pixels = phase_screens[0].shape[0]
+
+    phs2Rad = 2 * numpy.pi / (wavelength * 10 ** 9)
+
+    # Get initial up/down dependent params
+    if propagation_direction == "up":
+        ht = 0
+        ht_final = source_altitude
+        if ht_final==0:
+            raise ValueError("Can't propagate up to infinity")
+        scrnAlts = layer_altitudes
+
+        EFieldBuf = output_mask.copy().astype(CDTYPE)
+        logger.debug("Create EField Buf of mask")
+
+    else:
+        ht = layer_altitudes[scrnNo-1]
+        ht_final = 0
+        scrnRange = scrnRange[::-1]
+        scrnAlts = layer_altitudes[::-1]
+        EFieldBuf = numpy.exp(
+                1j*numpy.zeros((nx_output_pixels,) * 2)).astype(CDTYPE)
+        logger.debug("Create EField Buf of zero phase")
+
+    # Propagate to first phase screen (if not already there)
+    if ht!=scrnAlts[0]:
+        logger.debug("propagate to first phase screen")
+        z = abs(scrnAlts[0] - ht)
+        EFieldBuf[:] = opticalpropagation.angularSpectrum(
+                    EFieldBuf, wavelength,
+                    output_pixel_scale, output_pixel_scale, z)
+
+    # Go through and propagate between phase screens
+    for i in scrnRange:
+
+        phase = phase_screens[i]
+        # print("Got phase")
+
+        # Convert phase to radians
+        phase *= phs2Rad
+
+        # Change sign if propagating up
+        if propagation_direction == 'up':
+            phase *= -1
+        # print("Get distance")
+        # Get propagation distance for this layer
+        if i==(scrnNo-1):
+            z = abs(ht_final - ht) - z_total
+        else:
+            z = abs(scrnAlts[i+1] - scrnAlts[i])
+
+        # Update total distance counter
+        z_total += z
+
+        # print("Make EField")
+        # Apply phase to EField
+        EFieldBuf *= numpy.exp(1j*phase)
+
+        # Do ASP for last layer to next
+        EFieldBuf[:] = opticalpropagation.angularSpectrum(
+                    EFieldBuf, wavelength,
+                    output_pixel_scale, output_pixel_scale, z)
+
+        # logger.debug("Propagation: {}, {} m. Total: {}".format(i, z, z_total))
+
+    return EFieldBuf
